Remove QFT algorithm comment.
I originally reopened the file to add more comments, but
it occurred to me that it's probably better to not include
comments on the algorithm section at all - the criteria
require relatively precise interpretation, and the code
probably provides that as accurately and succinctly as
I could in words.

I will probably hate myself for this later.