Check QFT inputs for positivity, too.
This new bit of code checks that the QFT inputs are positive
(in addition to being numerics of identical length).  I also
updated the epsilon value to be called "tol", and used the
default value for tolerance used in the all.equal function.

If it's good enough for John Chambers, it's good enough
for me.