Move the verbosity logic into the master function.
S3 generics don't allow me to pass other arguments through to the subfuncs,
so I have to take care of this inside of the master fun.  Shouldn't be a
problem - I can't fathom there being enough QFTs in the world to render
the verbosity a memory problem.  gsub() and substr() should be damn fast,
too.