Pull objects from qft.obj, not environment; check that they exist.
Because I am an idiot, I had the criteria functions set up to pull
from the global environment instead of the qft.obj passed to them.
This fixes the cellestis.usa, and also adds a check for the correct variable
names in qft.obj.

This raises a question: should the data checking in qft.interp be abstracted
into its own function that's applied from within each of the qft.criteria
functions?
Pro: it's closer to the metal - users couldn't circumvent data cleaning by
working directly with the qft.criteria functions.
Con: they (and I!) have to remember to put it into their custom functions.