Acknowledge two failing tests.
Because of the way my tolerances work, they will sometimes provide
errors when the value is EXACTLY cutoff - tol.  For example, if the
tb - nil value is exactly 0.35 - tol, then the qft.criteria.cellestis.usa
function will:

1) add tol and then rightly declare that 0.35 is not greater than 0.35
2) add tol and then rightly declare that 0.35 is not less than 0.35

I can't imagine this ever happening, so I'm not worried about it - but
I'd prefer to leave the tests in.

Wait.  I'm a dumbass.  Just conver the positive test to >= 0.35.