If a criterion is a >=, make it >=
I had it in my head that tol would obviate the need for the = part of >=
entirely - and it almost does.  But when the value in a >= comparison is
EXACTLY cutoff - tol, then I still need that =.  This is silly and not
very practical because this realistically will probably never happen -
who generates QFT values with that much detail? - but now it's fixed.